Adjust tool execution list view
